chore: move matching logic to match.go and adding test cases
