+++ x/fundraising/types/match_test.go
@@ -1,0 +1,7 @@
+package types_test
+
+import "testing"
+
+func TestMatch(t *testing.T) {
+
+}
